Crate collada

Modules

document

Structs

Animation

BindData

BindDataSet

Geometry
A set of shapes, all using the given material.

Joint

JointWeights
Represents the weights of any joints that should control the vertex with skinned animation

ObjSet
A set of objects from the Collada document

Object
A mesh object.

Skeleton

TVertex
A single 2-dimensional point on a texture. "Texure Vertex".

Vertex
A single 3-dimensional point on the corner of an object.

VertexWeight

Enums

Shape
The various shapes supported by this library.

Constants

ROOT_JOINT_PARENT_INDEX

Functions

mat4_id

Type Definitions

JointIndex

Matrix4

Normal
A single 3-dimensional normal

NormalIndex
An index into the normals array of an object.

TextureIndex
An index into the texture vertex array of an object.

VTNIndex
An index into the vertex array, with an optional index into the texture array. This is used to define the corners of shapes which may or may not be textured.

VertexIndex
An index into the vertices array of an object, representing a vertex in the mesh. After parsing, this is guaranteed to be a valid index into the array, so unchecked indexing may be used.

WeightIndex