Struct collada::Animation

pub struct Animation {
    pub target: String,
    pub sample_times: Vec<f32>,
    pub sample_poses: Vec<Matrix4<f32>>,
}

A COLLADA animation consists of mapping of sample times to pose transforms for a single node in the scene (usually a skeleton joint)

Note - COLLADA supports animating arbitrary 'outputs', not just pose transforms, (eg colors, texture offsets, etc), but we'll leave those unsupported for now.

Fields

target

The node (joint) this animation is targeting

sample_times

Times for each sample (in seconds)

sample_poses

Node pose transforms for each sample. Column-major.