Struct collada::Object

pub struct Object {
    pub name: String,
    pub vertices: Vec<Vertex>,
    pub joint_weights: Vec<JointWeights>,
    pub tex_vertices: Vec<TVertex>,
    pub normals: Vec<Normal>,
    pub geometry: Vec<Geometry>,
}

A mesh object.

Fields

name

A human-readable name for this object. This can be set in blender.

vertices

The set of vertices this object is composed of. These are referenced by index in faces.

joint_weights

The set of attached joints for each vertex. Should match length of 'vertices' if present

tex_vertices

The set of texture vertices referenced by this object. The actual vertices are indexed by the second element in a VTNIndex.

normals

The set of normals referenced by this object. This are are referenced by the third element in a VTNIndex.

geometry

A set of shapes (with materials applied to them) of which this object is composed.